refactor(approvals): remove dead wrapApprovalTools (keep its coverage on the live path)
After the plan-09/10 consolidation, applyPolicy resolves each tool's AccessSpec
and calls wrapToolWithApproval directly — wrapApprovalTools (the old
ToolSet-level wrapper that read the approval() marker) has zero production
callers. A full-codebase audit (16 dimensions) independently rediscovered it as
the only genuinely-dead symbol the migration left behind.

- Remove it from production: the function + its now-orphaned getApprovalOptions
  and ToolSet imports + the index.ts export.
- Scrub the stale doc-comment mentions (policy/apply.ts, approvals/types.ts;
  also drops "filterAdmin", which never existed in this tree).
- KEEP runtime.test.ts — those tests exercise the LIVE wrapToolWithApproval (the
  human-in-the-loop approval gate); they just drove it through wrapApprovalTools.
  Re-point them via a small local test driver so wrapToolWithApproval stays fully
  covered. (Deleting the test would have dropped approvals/runtime.ts to 56%
  branch — the audit's "apply.test covers it" claim was over-optimistic.)

wrapToolWithApproval, getApprovalOptions, and the approval()/access() test
markers all stay. Coverage 99.27%; typecheck, lint, knip clean.
